+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 This is a history of changes to k13labs/futurama
 
+# 0.6.2
+* Change default async output to channel of size 1 to more easily support async merge and other ops
+* Add purpose-built `async-future` and `async-deferred` macros to more easily create either.
+
 # 0.6.1
 * Code linting updates